// libs/shinkai-node-state/src/v2/queries/getEncryptionKeys/types.ts
export type Encryptionkeys = {
  my_device_encryption_pk: string;
  my_device_encryption_sk: string;

  my_device_identity_pk: string;
  my_device_identity_sk: string;

  profile_encryption_pk: string;
  profile_encryption_sk: string;

  profile_identity_pk: string;
  profile_identity_sk: string;
};